+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### New features
 - Supports FlowConfig feature to set source to forward via HTTP request
+- Added 'Localhost' as interface to perform internal HTTP requests
 - Provide version of module via 'OnNewStatusModuleVersion'
 - Function 'getParameters' to provide PersistentData parameters
 - Check if features of module can be used on device and provide this via 'OnNewStatusModuleIsActive' event / 'getStatusModuleActive' function
@@ -18,6 +19,9 @@ All notable changes to this project will be documented in this file.
 - 'loadParameters' returns its success
 - 'sendParameters' can control if sent data should be saved directly by CSK_Module_PersistentData
 - Added UI icon and browser tab information
+
+### Bugfix
+- Error if module is not active but 'getInstancesAmount' was called
 
 ## Release 1.0.0
 - Initial commit